add user utils functions
