+++ src/server/workspace/src/auth/guards/local.guard.ts
@@ -1,6 +1,6 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { Observable } from 'rxjs';
+import { ExecutionContext, Injectable } from '@nestjs/common';
 
 @Injectable()
 export class LocalGuard extends AuthGuard('local') {
